requirements.txt
pbkdf2
psycopg2~=2.9.9
crypto
rich
textual
pyscrypt
httpx
argon2-cffi
